pyproject.toml
[tool.poetry]
name = "wagtail-sb-material"
version = "0.1.0"
description = ""
authors = ["zodiacfireworks <martin.vuelta@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
wagtail = "^2.13.4"
django-compressor = "^2.4.1"
django-settings-export = "^1.2.1"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
